# Cache parsed object names