Looking up the fstring table isn't that cheap, if we assume most object names are likely to re-appear, it makes sense to keep a small cache of them on the stack.

e.g. `[{"foo": 1, "bar": 2}, {"foo": 3, "bar": 4}]`

In term of implementation, we use a simple sorted array with binary search as it's the most compact and guarantee a decent `O(log n)`, and the comparison is first done on the string length, an then fallback to `memcmp`.

This helps quite a bit on `activitypub` and `twitter` benchmarks, however it is worse on `citm_catalog` because it has a number of unique object names, of the same size, so it behave poorly with the chosen sorting algorithm.

~~It's a bit of a double edge sword, but perhaps a smarter comparison function may solve it.~~ Solved later on.

Before:

```
== Parsing activitypub.json (58160 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json   714.000 i/100ms
                  oj   801.000 i/100ms
          Oj::Parser   943.000 i/100ms
           rapidjson   632.000 i/100ms
Calculating -------------------------------------
                json      7.135k (± 0.7%) i/s  (140.16 μs/i) -     35.700k in   5.003978s
                  oj      7.991k (± 0.2%) i/s  (125.14 μs/i) -     40.050k in   5.012044s
          Oj::Parser      9.611k (± 0.2%) i/s  (104.04 μs/i) -     48.093k in   5.003723s
           rapidjson      6.318k (± 0.2%) i/s  (158.29 μs/i) -     31.600k in   5.001896s

Comparison:
                json:     7134.7 i/s
          Oj::Parser:     9611.5 i/s - 1.35x  faster
                  oj:     7990.8 i/s - 1.12x  faster
           rapidjson:     6317.6 i/s - 1.13x  slower

== Parsing twitter.json (567916 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json    57.000 i/100ms
                  oj    62.000 i/100ms
          Oj::Parser    78.000 i/100ms
           rapidjson    56.000 i/100ms
Calculating -------------------------------------
                json    573.527 (± 1.6%) i/s    (1.74 ms/i) -      2.907k in   5.070094s
                  oj    619.368 (± 1.6%) i/s    (1.61 ms/i) -      3.100k in   5.006550s
          Oj::Parser    770.095 (± 0.9%) i/s    (1.30 ms/i) -      3.900k in   5.064768s
           rapidjson    560.601 (± 0.4%) i/s    (1.78 ms/i) -      2.856k in   5.094597s

Comparison:
                json:      573.5 i/s
          Oj::Parser:      770.1 i/s - 1.34x  faster
                  oj:      619.4 i/s - 1.08x  faster
           rapidjson:      560.6 i/s - 1.02x  slower

== Parsing citm_catalog.json (1727030 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json    31.000 i/100ms
                  oj    34.000 i/100ms
          Oj::Parser    46.000 i/100ms
           rapidjson    38.000 i/100ms
Calculating -------------------------------------
                json    319.842 (± 0.3%) i/s    (3.13 ms/i) -      1.612k in   5.040026s
                  oj    329.315 (± 2.4%) i/s    (3.04 ms/i) -      1.666k in   5.061887s
          Oj::Parser    452.725 (± 1.1%) i/s    (2.21 ms/i) -      2.300k in   5.080996s
           rapidjson    358.160 (± 0.8%) i/s    (2.79 ms/i) -      1.824k in   5.093054s

Comparison:
                json:      319.8 i/s
          Oj::Parser:      452.7 i/s - 1.42x  faster
           rapidjson:      358.2 i/s - 1.12x  faster
                  oj:      329.3 i/s - 1.03x  faster

```

After:

```
== Parsing activitypub.json (58160 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json   810.000 i/100ms
                  oj   794.000 i/100ms
          Oj::Parser   936.000 i/100ms
           rapidjson   630.000 i/100ms
Calculating -------------------------------------
                json      7.759k (± 1.2%) i/s  (128.88 μs/i) -     38.880k in   5.011657s
                  oj      7.945k (± 1.2%) i/s  (125.87 μs/i) -     40.494k in   5.097765s
          Oj::Parser      9.658k (± 1.3%) i/s  (103.54 μs/i) -     48.672k in   5.040614s
           rapidjson      6.321k (± 0.3%) i/s  (158.20 μs/i) -     32.130k in   5.082985s

Comparison:
                json:     7759.2 i/s
          Oj::Parser:     9657.8 i/s - 1.24x  faster
                  oj:     7944.6 i/s - same-ish: difference falls within error
           rapidjson:     6321.1 i/s - 1.23x  slower

== Parsing twitter.json (567916 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json    66.000 i/100ms
                  oj    62.000 i/100ms
          Oj::Parser    78.000 i/100ms
           rapidjson    56.000 i/100ms
Calculating -------------------------------------
                json    661.653 (± 0.6%) i/s    (1.51 ms/i) -      3.366k in   5.087447s
                  oj    621.674 (± 1.3%) i/s    (1.61 ms/i) -      3.162k in   5.087171s
          Oj::Parser    775.859 (± 0.3%) i/s    (1.29 ms/i) -      3.900k in   5.026718s
           rapidjson    563.961 (± 0.4%) i/s    (1.77 ms/i) -      2.856k in   5.064230s

Comparison:
                json:      661.7 i/s
          Oj::Parser:      775.9 i/s - 1.17x  faster
                  oj:      621.7 i/s - 1.06x  slower
           rapidjson:      564.0 i/s - 1.17x  slower

== Parsing citm_catalog.json (1727030 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json    30.000 i/100ms
                  oj    34.000 i/100ms
          Oj::Parser    47.000 i/100ms
           rapidjson    38.000 i/100ms
Calculating -------------------------------------
                json    303.318 (± 0.3%) i/s    (3.30 ms/i) -      1.530k in   5.044285s
                  oj    328.832 (± 2.7%) i/s    (3.04 ms/i) -      1.666k in   5.069899s
          Oj::Parser    458.235 (± 0.4%) i/s    (2.18 ms/i) -      2.303k in   5.025895s
           rapidjson    381.204 (± 0.5%) i/s    (2.62 ms/i) -      1.938k in   5.084065s

Comparison:
                json:      303.3 i/s
          Oj::Parser:      458.2 i/s - 1.51x  faster
           rapidjson:      381.2 i/s - 1.26x  faster
                  oj:      328.8 i/s - 1.08x  faster

```