@byroot 2024-11-02
https://github.com/ruby/json/pull/675/commits/4a04252023da41facf3ee0cc7f80c2c91b29f3e0 solves the regression.

It's simple, we just don't cache names that don't start with a letter. This may sound like benchmark gaming, but I think it makes sense as an heuristic given how most JSON documents are constructed.

```
== Parsing activitypub.json (58160 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json   779.000 i/100ms
                  oj   793.000 i/100ms
          Oj::Parser   956.000 i/100ms
           rapidjson   620.000 i/100ms
Calculating -------------------------------------
                json      7.849k (± 0.3%) i/s  (127.41 μs/i) -     39.729k in   5.061875s
                  oj      7.953k (± 0.3%) i/s  (125.74 μs/i) -     40.443k in   5.085293s
          Oj::Parser      9.652k (± 0.2%) i/s  (103.61 μs/i) -     48.756k in   5.051573s
           rapidjson      6.198k (± 0.1%) i/s  (161.35 μs/i) -     31.000k in   5.001787s

Comparison:
                json:     7848.7 i/s
          Oj::Parser:     9651.7 i/s - 1.23x  faster
                  oj:     7953.0 i/s - 1.01x  faster
           rapidjson:     6197.8 i/s - 1.27x  slower


== Parsing twitter.json (567916 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json    66.000 i/100ms
                  oj    62.000 i/100ms
          Oj::Parser    79.000 i/100ms
           rapidjson    56.000 i/100ms
Calculating -------------------------------------
                json    676.145 (± 0.3%) i/s    (1.48 ms/i) -      3.432k in   5.075895s
                  oj    631.864 (± 0.3%) i/s    (1.58 ms/i) -      3.162k in   5.004287s
          Oj::Parser    791.409 (± 0.3%) i/s    (1.26 ms/i) -      4.029k in   5.090947s
           rapidjson    566.597 (± 0.2%) i/s    (1.76 ms/i) -      2.856k in   5.040641s

Comparison:
                json:      676.1 i/s
          Oj::Parser:      791.4 i/s - 1.17x  faster
                  oj:      631.9 i/s - 1.07x  slower
           rapidjson:      566.6 i/s - 1.19x  slower


== Parsing citm_catalog.json (1727030 bytes)
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                json    34.000 i/100ms
                  oj    32.000 i/100ms
          Oj::Parser    43.000 i/100ms
           rapidjson    37.000 i/100ms
Calculating -------------------------------------
                json    356.414 (± 0.3%) i/s    (2.81 ms/i) -      1.802k in   5.055963s
                  oj    349.397 (± 2.0%) i/s    (2.86 ms/i) -      1.760k in   5.039216s
          Oj::Parser    478.342 (± 0.4%) i/s    (2.09 ms/i) -      2.408k in   5.034158s
           rapidjson    395.222 (± 0.8%) i/s    (2.53 ms/i) -      1.998k in   5.055669s

Comparison:
                json:      356.4 i/s
          Oj::Parser:      478.3 i/s - 1.34x  faster
           rapidjson:      395.2 i/s - 1.11x  faster
                  oj:      349.4 i/s - same-ish: difference falls within error
```