@byroot 2024-11-02
@eregon while you're here, I'm surprised to see TruffleRuby is missing `rb_enc_interned_str`:

```
/home/runner/.rubies/truffleruby-head/bin/truffleruby: symbol lookup error: /home/runner/work/json/json/ext/json/ext/parser.so: undefined symbol: rb_enc_interned_str
```