@eregon 2024-11-02
https://github.com/oracle/truffleruby/blob/07b978e4e0d43827ca13c600af447c98ac477a71/src/main/c/cext/string.c#L435-L442 looks like TruffleRuby has rb_enc_interned_str_cstr() and rb_str_to_interned_str() but not rb_enc_interned_str() yet -> https://github.com/oracle/truffleruby/issues/3703